Remove signature confirmation and cash/bank split logic
This commit removes all code and dependencies related to signature-based confirmation, including UI components, hooks, and the 'react-signature-canvas' package. It also simplifies order and payment display logic by removing the distinction between cash and bank payments, showing only the total sum. Translation keys and related logic for signature and payment split have been cleaned up for all supported languages.
