+++ package.json
@@ -31,7 +31,6 @@
     "react-icons": "^5.5.0",
     "react-media-devices": "^1.1.5",
     "react-router-dom": "^7.0.2",
-    "react-signature-canvas": "^1.1.0-alpha.2",
     "react-zxing": "^2.0.2",
     "swiper": "^11.1.15"
   },
