+++ src/components/order page components/SameClientsOrders.tsx
@@ -4,7 +4,6 @@ interface SameClientsOrdersProps {
   selectedOrdersList: {
     tracking_code: string;
     sum: number;
-    sumcash: number;
     client_address: string;
     places?: any[];
     parcel_with_return?: boolean;
@@ -76,22 +75,7 @@ if (showReturnParcels) {
               </div>
               <div className="flex justify-between items-center">
                 <span className="font-base text-sm">{t("sum")} :</span>
-                { order.sumcash == order.sum ? (
-                  <span className="font-bold text-sm text-blue-700">{order.sum} ₾</span>
-                ) : (
-                <div>
-                  <div className="flex gap-2">
-                    <span className="font-bold text-sm text-blue-700">{t("Cash2")}</span>
-                    <span className="font-bold text-sm text-blue-700">{order.sumcash} ₾</span>
-                  </div>
-                  <div className="flex gap-2">
-                    <span className="font-bold text-sm text-blue-700">{t("Bank")}</span>
-                    <span className="font-bold text-sm text-blue-700">{order.sum} ₾</span>
-                  </div>
-                </div>
-                )
-
-                }
+                <span className="font-base text-sm">{order.sum} ₾</span>
               </div>
             </div>
           </li>
