+++ src/components/returnDeclineModal.tsx
@@ -89,10 +89,7 @@ const ReturnDeclineModal: React.FC<ReturnDeclineModalProps> = ({
           >
             {t("Cancel")}
           </button>
-          <button
-            onClick={handleconfirm}
-            className={`px-4 py-2 text-black font-semibold rounded-md `}
-          >
+          <button onClick={handleconfirm} className={`px-4 py-2 text-black font-semibold rounded-md `}>
             {t("Confirm")}
           </button>
         </div>
