+++ src/pages/OrderPage.tsx
@@ -27,7 +27,6 @@ const OrderPage = () => {
     setSelectedOrders,
     selectedOrders,
     totalSum,
-    totalSumCash,
     totalQuantity,
     handleCheckboxChange,
   } = useOrder(selectedOrdersList);
@@ -197,32 +196,9 @@ const OrderPage = () => {
               </div>
             </div>
             {(order.Status !== "Accepted" || order?.places) && (
-              <div className="p-1 flex justify-between items-center">
+              <div className="p-1 flex justify-between">
                 <span className="font-base text-sm">{t("sum")} :</span>
-                {order.sumcash == order.sum ? (
-                  <span className="font-bold text-sm text-blue-700">
-                    {order.sum} ₾
-                  </span>
-                ) : (
-                  <div>
-                    <div className="flex gap-2">
-                      <span className="font-bold text-sm text-blue-700">
-                        {t("Cash2")}
-                      </span>
-                      <span className="font-bold text-sm text-blue-700">
-                        {order.sumcash} ₾
-                      </span>
-                    </div>
-                    <div className="flex gap-2">
-                      <span className="font-bold text-sm text-blue-700">
-                        {t("Bank")}
-                      </span>
-                      <span className="font-bold text-sm text-blue-700">
-                        {order.sum} ₾
-                      </span>
-                    </div>
-                  </div>
-                )}
+                <span className="font-base text-sm">{order.sum} ₾</span>
               </div>
             )}
             <div className="p-1 flex justify-between">
@@ -278,22 +254,9 @@ const OrderPage = () => {
                   <span>{t("Total quantity")} :</span>
                   <span className="font-medium">{totalQuantity}</span>
                 </div>
-                <div className="flex justify-between items-center">
+                <div className="p-1 flex justify-between">
                   <span>{t("Total amount")} :</span>
-                  {totalSumCash === totalSum ? (
-                    <span className="font-medium text-s text-blue-700">
-                      {totalSum} ₾
-                    </span>
-                  ) : (
-                    <div className="flex flex-col">
-                      <span className="font-medium text-xs text-blue-700">
-                        {t("Cash2")} {totalSumCash} ₾
-                      </span>
-                      <span className="font-medium text-xs text-blue-700">
-                        {t("Bank")} {totalSum} ₾
-                      </span>
-                    </div>
-                  )}
+                  <span className="font-medium">{totalSum} ₾</span>
                 </div>
                 <div className="flex space-x-4">
                   <Button
